tests/test_lambda_function.py
import unittest
from unittest.mock import patch, Mock
import json
import os
from your_lambda_file import (  # Replace 'your_lambda_file' with the actual file name
    get_bbox_and_coords_from_geojson,
    clipper,
    write_tiff_and_upload,
    calculate_data,
    get_next_execution_name,
    lambda_handler
)

class TestEnvironmentVariables(unittest.TestCase):

    @patch.dict(os.environ, {'STAC_API_ENDPOINT': 'https://example.com'})
    def test_stac_api_endpoint(self, mock_env):
        self.assertEqual(os.environ.get('STAC_API_ENDPOINT'), 'https://example.com')


class TestAWSClients(unittest.TestCase):

    @patch('boto3.client')
    def test_s3_client(self, mock_client):
        self.assertEqual(s3, mock_client('s3'))

    @patch('boto3.client')
    def test_sns_client(self, mock_client):
        self.assertEqual(sns, mock_client('sns'))

    @patch('boto3.client')
    def test_sfn_client(self, mock_client):
        self.assertEqual(sfn, mock_client('stepfunctions'))


class TestGetBboxAndCoordsFromGeojson(unittest.TestCase):

    def test_valid_geojson(self):
        geojson_str = '{"geometry": {"coordinates": [[1, 2], [3, 4]]}}'
        bbox, coords = get_bbox_and_coords_from_geojson(geojson_str)
        self.assertEqual(bbox, [1, 2, 3, 4])
        self.assertEqual(coords, [[1, 2], [3, 4]])


class TestClipperFunction(unittest.TestCase):

    @patch('requests.get')
    @patch('rasterio.open')
    def test_clipper(self, mock_rasterio_open, mock_requests_get):
        mock_requests_get.return_value.content = b'some_content'
        mock_rasterio_open.return_value.__enter__.return_value.read.return_value = 'some_data'
        # Add more mock setups as needed
        result = clipper('some_url', 'some_shapes', 'some_clipped_band')
        self.assertEqual(result, 'some_expected_result')  # Replace with the actual expected result


class TestWriteTiffAndUpload(unittest.TestCase):

    @patch('s3.upload_file')
    @patch('rasterio.open')
    def test_write_tiff_and_upload(self, mock_rasterio_open, mock_s3_upload):
        # Add mock setups and test the function
        pass


class TestCalculateData(unittest.TestCase):

    def test_calculate_data(self):
        # Add mock setups and test the function
        pass


class TestGetNextExecutionName(unittest.TestCase):

    @patch('sfn.describe_execution')
    def test_get_next_execution_name(self, mock_describe_execution):
        # Add mock setups and test the function
        pass


class TestLambdaHandler(unittest.TestCase):

    @patch('s3.get_object')
    @patch('sns.publish')
    @patch('sfn.start_execution')
    def test_lambda_handler(self, mock_start_execution, mock_sns_publish, mock_s3_get_object):
        event = {'Records': [{'s3': {'bucket': {'name': 'some_bucket'}, 'object': {'key': 'some_key'}}}]}
        context = Mock()
        # Add more mock setups as needed
        result = lambda_handler(event, context)
        self.assertEqual(result, 'some_expected_result')  # Replace with the actual expected result


if __name__ == '__main__':
    unittest.main()
